Inline take/dropWhile functions after first argument application
takeWhile was a hotspot in a ticky profile of our application code. I
then noticed in two microbenchmarks:

    bench "takeWhile" $ nf (\t -> T.takeWhile (/= '2') t) someText

         449,707,656 bytes allocated in the heap   for 10000000 iterations
        105.4 ns   (105.0 ns .. 106.5 ns) per iteration

    bench "takeWhile" $ nf (T.takeWhile (/= '2')) someText

     114,847,742,304 bytes allocated in the heap  for 10000000 iterations
        2.963 μs   (2.721 μs .. 3.176 μs) per iteration

The allocations in the latter is surprising; I'm not sure why we need to
inline to get this benefit. In the former case we allocate ~32bytes per
iteration (the Text constructor, presumably); it should be possible to
get that from optimizing Data.Text in isolation AFAICT...

Inlining after the first application should benefit attoparsec:
https://github.com/haskell/attoparsec/blob/master/Data/Attoparsec/Text/Internal.hs#L240

I also removed the phase annotations on the INLINE pragma.
It might be worth doing that to the other functions that no longer have
fusion RULEs. It also might be good to do this same eta fiddling for
other inlined functions.